Add utils for tag drop rate numbers
