+++ src/utils.py
@@ -1,5 +1,7 @@
 from pathlib import Path
 
+from src.tag_utils import get_tag_drop_rate_str
+
 TIMEOUT_IN_SECONDS = 5
 
 
@@ -36,8 +38,6 @@ def get_listing_output_file_name_suffix(
     tag_drop_rate_str: str | None = None,
     rarity: str | None = None,
 ) -> str:
-    from src.market_search import get_tag_drop_rate_str
-
     if tag_drop_rate_str is None:
         tag_drop_rate_str = get_tag_drop_rate_str(rarity=rarity)
